Add api - `into_stdio` .
